Small data adjustment (#9)
Changed the detection range to the nearest centimeter rather than a
double. This slightly helps conserve data in the long run as the double
was unnecessary. In addition, the calibration has been reintroduced as
it turns out the calibration was not broken to begin with. Lastly, a few
functions / setup procedures have been renamed to increase clarity. This
is a very small change and should not affect current performance.