Created file HardwareProperties.h to keep track of the values of the pins on the board.
Created class AbstractSensor to try and encapsulate the functions for reading
data from either an analog pin or digital pin.

The #include preprocessor tokens for the "arduino.h" header file were changed to "Arduino.h" because that is what it
is named on my system. It should be changed back if it is causing compilation
errors.