Object detection (#14)
* Object Distance

First commit to the Object distance module. The next thing to work on
would be the vehicle being able to reason about an object being in its
blind spot. Currently it reads the most accurate values of the two.

* Object Blind Spot Logic

This change includes logic for the blind spot detection of the robot.
The next change list will handle the sensor control detection and will
implement a test code.

* Removed collision checks

Removed old collision checks as they will be implemented later through a
new module

* Object Detection tasks

This change adds additional checks to the object in blind spot logic
currently implemented on the robot and adds additonal tasks that need to
be completed within the object detection module.

* Bug noting and some fixes

This fixes some bugs I didn't know I had. I removed calibration again;
it isn't working very well at the moment @_@. Added TODOs taht need
improvement

* Small Modifications

Tweaked the logic in the object read and sensor over lap exists boolean.
The next change list is going to include a revamped blind spot logic.

* Blindspot check revamp

This change includes a revamp of the blind spot logic. In addition it
also includes some changes to the computational logic to help speed up
calculation between itterations. This remains to be tested which will be
done in the near future.

* Changed order of declaration

* Approaching object speed (#12)

* Approaching object speed

Intial commit for the approaching object sped logic.

* Util adjustments

This change reformats the functions found within util.cpp to a class.
This should improve calcuation as we only run this once, rather than
calling it each time we want to analyze an input array. Should maintain
same functionality as before. *NOTE* This needs to be tested.

* Revamps array analysis

This change revamps array analysis and adds the object speed detection.
With this change the robot can detect the speed of the approaching
object.

* Object Status

This change includes a simple evaluation on the approxiate object
maneuver. IE, if it is approaching, leaving or is static.