Rearranged location of some functions and added Pluggable interface.
The pluggable interface is to ensure that every piece of hardware that can be
pluplugged into the board has an attach method. Moved the hardware_begin method
into the hardware properties file. The hardware properties file seems to be the
location where all functions handeling setup and wiring with the arduino board
should be located.